Обновляет код модели User
